moneybot.testing
Moved MarketHistoryMock there for now
